requirements.txt
pytest
allure-pytest
selene==2.0.0rc9
appium-python-client
python-dotenv
pydantic-settings
requests